added option to disable ssl verification
When not using autodiscovery, chances are, that the server has a
self-signed certificate. In that case, we may want to disable ssl
verification. Assuming, that we configured server_url, we know what we
are doing and consciously accept the risk. Default in config file is
to verify.

Please notice, that newer version of exchangelibs (>1.10) uses
NoVerifyHTTPAdapter for the same purpose. Then the code should utilise
this:

  from exchangelib.protocol import NoVerifyHTTPAdapter
  BaseProtocol.HTTP_ADAPTER_CLS = NoVerifyHTTPAdapter

Unfortunatelly, with older exchangelibs, we have to uese verif_ssl.